Moved mobile ui widget to own file
